Refactored the files into corresponding folders
