--- src/pages/ErrorPage.tsx
@@ -1,5 +1,0 @@
-const ErrorPage = () => {
-	return <div>ErrorPage</div>;
-};
-
-export default ErrorPage;
