+++ src/pages/Registration/Patient/PatientRegistrationPage.tsx
